requirements.txt
dashscope
pyaudio
python-dotenv
pillow
mss
